requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
botasaurus==3.2.27
botasaurus-proxy-authentication==1.0.12
capsolver-extension-python==1.0.1
casefy==0.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
chrome-extension-python==1.0.1
chromedriver-autoinstaller-fix==1.0.6
cloudscraper==1.2.71
dateutils==0.6.12
et-xmlfile==1.1.0
frozenlist==1.4.1
geonamescache==2.0.0
h11==0.14.0
idna==3.6
javascript-fixes==1.1.18
joblib==1.3.2
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
psutil==5.9.8
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==1.26.18
wsproto==1.2.0
yarl==1.9.4
